fix throw exception when not found

+++ src/main/java/com/pii/tracker/track_calories/user/service/UserService.java
@@ -46,6 +46,9 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("Пользователь не найден id=" + id);
+        }
         userRepository.deleteById(id);
     }
 }
